refactor(SingleTrackPlayer): Moved to core refactor(MatomoUtils): Moved to core as required by the player, UI stuff kept in an extension in main app
